requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2025.2
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.5
gunicorn==21.2.0 